+++ matrix_actions.cpp
@@ -1,10 +1,9 @@
 #include "matrix_actions.h"
 
-#include <cstddef>
 #include <stdexcept>
 
 
-void transpose(double* matrix, const int& rows, const int& cols) {
+void transpose(double* matrix, const size_t& rows, const size_t& cols) {
     double tmatrix[1000]{0.0};
     for (size_t i = 0ull; i < rows; ++i) {
         for (size_t j = 0ull; j < cols; ++j) {
@@ -16,7 +15,7 @@ void transpose(double* matrix, const int& rows, const int& cols) {
     }
 }
 
-void oddEvenRows(double* matrix, const int& rows, const int& cols) {
+void oddEvenRows(double* matrix, const size_t& rows, const size_t& cols) {
     for (size_t j = 0ull; j < cols; ++j) {
         for (size_t i = 0ull; i < (rows - rows % 2); i += 2) {
             double t = matrix[i * cols + j];
