+++ matrixio.cpp
@@ -1,10 +1,9 @@
 #include "matrixio.h"
 
-#include <cstddef>
 #include <iostream>
 
 
-void matrixin(double * matrix, const int& rows, const int& cols) {
+void matrixin(double * matrix, const size_t& rows, const size_t& cols) {
     for (size_t i = 0ull; i < rows; ++i) {
         for (size_t j = 0ull; j < cols; ++j) {
             std::cin >> matrix[i * cols + j];
@@ -15,7 +14,7 @@ void matrixin(double * matrix, const int& rows, const int& cols) {
     }
 }
 
-void matrixout(double * matrix, const  int& rows, const int& cols) {
+void matrixout(double * matrix, const  size_t& rows, const size_t& cols) {
     for (size_t i = 0ull; i < rows; ++i) {
         std::cout << matrix[i * cols];
         for (size_t j = 1ull; j < cols; ++j) {
